Forward declare tree_code in coretypes.h
The following changes the reassociation_width target hook to take
a tree_code after recognizing the x86 hook compares against rtx_code PLUS.
Like rtx_code we can now forward declare tree_code to achieve that.

gcc/
	* coretypes.h (enum tree_code): Declare.
	* tree-core.h (enum tree_code): Adjust.
	* target.def (reassociation_width): Get a tree_code instead
	of unsigned int.
	* doc/tm.texi: Regenerate.
	* genmatch.cc (enum tree_code): Adjust.
	* gimple-loop-versioning.cc (address_info_hasher::hash): Likewise.
	* config/aarch64/aarch64.cc (aarch64_reassociation_width):
	Likewise.
	* config/loongarch/loongarch.cc
	(loongarch_cpu_sched_reassociation_width): Likewise.
	* config/mips/mips.cc (mips_sched_reassociation_width): Likewise.
	* config/rs6000/rs6000.cc (rs6000_reassociation_width): Likewise.
	* config/i386/i386.cc (ix86_reassociation_width): Likewise
	and fix bogus compares.

gcc/cp/
	* module.cc (trees_out::decl_value): Adjust for enum class
	tree_code.

+++ gcc/gimple-loop-versioning.cc
@@ -449,7 +449,7 @@ hashval_t
 address_info_hasher::hash (const address_info *info)
 {
   inchash::hash hash;
-  hash.add_int (info->base ? TREE_CODE (info->base) : 0);
+  hash.add_int (info->base ? (unsigned) TREE_CODE (info->base) : 0);
   hash.add_int (info->terms.length ());
   for (unsigned int i = 0; i < info->terms.length (); ++i)
     {
